Refactor redux config
