+++ frontend/src/App.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 
-import { store } from './app/store';
+import { store } from './store/store';
 import { App } from './App';
 
 test('renders learn react link', () => {
